requirements.txt

alembic==1.8.1
appnope @ file:///opt/concourse/worker/volumes/live/5f13e5b3-5355-4541-5fc3-f08850c73cf9/volume/appnope_1606859448618/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
cycler==0.11.0
Cython==0.29.32
databricks-cli==0.17.3
debugpy @ file:///opt/concourse/worker/volumes/live/e17abc9f-08a9-4dca-4e9b-e6d24bde47ce/volume/debugpy_1637091828617/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
distlib==0.3.6
docker==5.0.3
efficientnet-pytorch==0.7.1
entrypoints @ file:///opt/concourse/worker/volumes/live/84dd3e87-c86b-405f-5113-9d1bdbfd10b9/volume/entrypoints_1649926452122/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.8.0
Flask==2.2.2
fonttools==4.37.1
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.3
gunicorn==20.1.0
identify==2.5.5
idna==3.3
imageio==2.21.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel @ file:///opt/concourse/worker/volumes/live/675cc243-e35d-4e85-47e2-838adcdd7698/volume/ipykernel_1647000791373/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_651b0309-eccd-4278-bb05-da83352f345cw81ya025/croots/recipe/ipython_1657652219297/work
itsdangerous==2.1.2
jedi @ file:///opt/concourse/worker/volumes/live/fb2e45f1-1dff-42fa-6cab-bd094a4c4372/volume/jedi_1644315264796/work
Jinja2==3.1.2
jupyter-core @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_8ba802e0-15f6-4a45-a9ee-144034526783uwxtvvew/croots/recipe/jupyter_core_1651671235924/work
jupyter_client @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_41tzpfqkok/croots/recipe/jupyter_client_1661848920196/work
kiwisolver==1.4.4
Mako==1.2.2
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mlflow==1.28.0
munch==2.5.0
nest-asyncio @ file:///opt/concourse/worker/volumes/live/a7f89d20-060a-480e-487d-25959e806725/volume/nest-asyncio_1649847918874/work
networkx==2.8.6
nodeenv==1.7.0
numpy==1.23.2
oauthlib==3.2.0
opencv-python-headless==4.5.2.52
packaging==21.3
pandas==1.4.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
pretrainedmodels==0.7.4
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==4.21.5
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@0d53acbcf5123d4c88040fe68fbb9805fc5b2fb9
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT==2.4.0
pyngrok==5.1.0
pyparsing==3.0.9
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_15f7a459-ad98-422b-b8da-cbf1f626e2115nt0ocwy/croots/recipe/pyzmq_1657724193704/work
querystring-parser==1.2.4
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
segmentation-models-pytorch==0.3.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tabulate==0.8.10
tifffile==2022.8.12
timm==0.4.12
toml==0.10.2
torch==1.12.1
torchvision==0.13.1
tornado @ file:///opt/concourse/worker/volumes/live/05341796-4198-4ded-4a9a-332fde3cdfd1/volume/tornado_1606942323372/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.5
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websocket-client==1.4.0
Werkzeug==2.2.2
zipp==3.8.1
